requirements.txt
# test
pytest
coverage>=5
pytest-cov
pylint
black
pre-commit

# run
falcon
pysmb
requests